+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 ## Version 1.0.0 for Minecraft 1.16.3
 
 *Should work with Forge 1.16.1 and later*
+
+- Shows the number of anvil uses
+- Shows the base repair cost
+- Works with any minecraft-supported language (english, russian, japanese, ...)